[TASK] cleanup code and implement new device handling
