+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -5,69 +5,14 @@ name: Bug report about: Create a report to let us know of any issues you encount
 
 ### Prerequisites
 
-* [ ] Model name of your Devices from Zigbee2Mqtt
+* [ ] Model name of your Devices
 * [ ] Output from Home Assistant Developer Tools state e.g.
+* [ ] Output from Home Assistant Device Diagnostic from BT
 
-```yaml
-hvac_modes:
-  - auto
-  - 'off'
-  - cool
-  - heat
-  - dry
-  - fan_only
-min_temp: 5
-max_temp: 35
-target_temp_step: 0.5
-preset_modes:
-  - none
-  - programming
-  - manual
-  - temporary_manual
-  - holiday
-current_temperature: 24
-temperature: 0
-preset_mode: manual
-battery: 94
-boost_heating: 'OFF'
-boost_heating_countdown: 0
-boost_heating_countdown_time_set: 300
-child_lock: UNLOCK
-current_heating_setpoint: 0
-device:
-  applicationVersion: 64
-  dateCode: ''
-  friendlyName: Demo - TVR
-  hardwareVersion: 1
-  ieeeAddr: '0x84fd27fffea65a1d'
-  manufacturerID: 4098
-  manufacturerName: _TZE200_b6wax7g0
-  model: BRT-100-TRV
-  networkAddress: 26598
-  powerSource: Battery
-  stackVersion: 0
-  type: EndDevice
-  zclVersion: 3
-eco_mode: 'OFF'
-eco_temperature: 22
-linkquality: 162
-local_temperature: 24
-local_temperature_calibration: 0
-max_temperature: 30
-min_temperature: 5
-position: 0
-preset: manual
-program_saturday:
-  saturday: '6h:0m 24°C,  12h:0m 23°C,   14h:30m 22°C,  17h:30m 21°C '
-program_sunday:
-  sunday: '  6h:0m 19°C,  12h:30m 20°C,  14h:30m 21°C,  18h:30m 20°C '
-program_weekday:
-  weekday: ' 6h:0m 20°C,  11h:30m 21°C,  13h:30m 21°C,  17h:30m 23°C '
-update_available: null
-window: CLOSED
-window_detection: 'OFF'
-friendly_name: Demo - TVR
-supported_features: 17
+```json
+{
+  YOUR DEVICE DIAGNOSTICS JSON OUTPUT HERE
+}
 ```
 
 ### Description
@@ -90,7 +35,7 @@ supported_features: 17
 
 ### Versions
 
-<!-- Provide both, HA and Z2M version -->
+<!-- Provide both, HA and BT version -->
 
 ### Additional Information
