+++ custom_components/better_thermostat/__init__.py
@@ -1,11 +1,12 @@
 """The better_thermostat component."""
-from homeassistant.const import Platform
+from homeassistant.const import Platform, EVENT_HOMEASSISTANT_STOP
 from homeassistant.core import HomeAssistant, Config
 from homeassistant.config_entries import ConfigEntry
-
 DOMAIN = "better_thermostat"
 PLATFORMS = [Platform.CLIMATE]
+import logging
 
+_LOGGER = logging.getLogger(__name__)
 
 async def async_setup(hass: HomeAssistant, config: Config):
     """Set up this integration using YAML is not supported."""
@@ -14,4 +15,19 @@ async def async_setup(hass: HomeAssistant, config: Config):
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     hass.config_entries.async_setup_platforms(entry, PLATFORMS)
+    entry.async_on_unload(entry.add_update_listener(config_entry_update_listener))
     return True
+
+async def config_entry_update_listener(hass, entry) -> None:
+    """Handle options update."""
+    await async_unload_entry(hass, entry)
+    await async_setup_entry(hass, entry)
+
+async def async_unload_entry(hass, entry):
+    """Unload a config entry."""
+    await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
+    return True
+
+async def async_reload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> None:
+    await async_unload_entry(hass, config_entry)
+    await async_setup_entry(hass, config_entry)
