+++ custom_components/better_thermostat/const.py
@@ -35,6 +35,10 @@ CONF_MIN_TEMP = "min_temp"
 CONF_MAX_TEMP = "max_temp"
 CONF_PRECISION = "precision"
 CONF_LOCAL_CALIBRATION = "local_calibration"
+CONF_CHILD_LOCK = "child_lock"
+CONF_CALIBRATIION_ROUND = "calibration_round"
+CONF_HEAT_AUTO_SWAPPED = "heat_auto_swapped"
+CONF_MODEL = "model"
 
 SUPPORT_FLAGS = SUPPORT_TARGET_TEMPERATURE
