+++ custom_components/better_thermostat/controlling.py
@@ -1,5 +1,6 @@
 import asyncio
 import logging
+from .events.trv import convert_outbound_states
 
 from homeassistant.components.climate.const import (
     HVAC_MODE_HEAT,
@@ -9,8 +10,7 @@ from homeassistant.components.climate.const import (
 from homeassistant.components.number.const import SERVICE_SET_VALUE
 from homeassistant.const import ATTR_TEMPERATURE
 
-from .models.models import convert_outbound_states
-from .models.utils import convert_to_float, round_to_hundredth_degree
+from .helpers import convert_to_float, round_to_hundredth_degree
 from .weather import check_weather
 
 _LOGGER = logging.getLogger(__name__)
@@ -294,7 +294,7 @@ async def trv_valve_maintenance(self):
         # get current HVAC mode from HA
         try:
             _last_hvac_mode = self.hass.states.get(self.heater_entity_id).state
-        except ValueError:
+        except:
             _LOGGER.error(
                 "better_thermostat %s: Could not load current HVAC mode", self.name
             )
@@ -393,7 +393,7 @@ async def trv_valve_maintenance(self):
         try:
             self.hass.states.get(self.heater_entity_id).attributes.get("valve_position")
             valve_position_available = True
-        except ValueError:
+        except:
             pass
 
         if valve_position_available:
