+++ custom_components/better_thermostat/helpers.py
@@ -4,10 +4,10 @@ import asyncio
 import logging
 import numbers
 from datetime import datetime
+from typing import Union
+from homeassistant.helpers import device_registry as dr, entity_registry as er
 
-from homeassistant.components.climate.const import HVAC_MODE_OFF
-from homeassistant.const import ATTR_TEMPERATURE, STATE_UNAVAILABLE, STATE_UNKNOWN
-
+from homeassistant.components.climate.const import HVAC_MODE_AUTO, HVAC_MODE_HEAT, HVAC_MODE_OFF
 from .const import (
     ATTR_STATE_CALL_FOR_HEAT,
     ATTR_STATE_DAY_SET_TEMP,
@@ -15,7 +15,6 @@ from .const import (
     ATTR_STATE_NIGHT_MODE,
     ATTR_STATE_WINDOW_OPEN,
 )
-from .models.utils import convert_to_float
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -44,8 +43,39 @@ def log_info(self, message):
     )
 
 
-async def startup(self):
-    """Run startup tasks.
+def mode_remap(self, hvac_mode: str, inbound: bool = False) -> str:
+    """Remap HVAC mode to correct mode if nessesary.
+
+    Parameters
+    ----------
+    self :
+            FIXME
+    hvac_mode : str
+            HVAC mode to be remapped
+
+    inbound : bool
+            True if the mode is coming from the device, False if it is coming from the HA.
+
+    Returns
+    -------
+    str
+            remapped mode according to device's quirks
+    """
+    _heat_auto_swapped = self._config.get("heat_auto_swapped")
+    if _heat_auto_swapped:
+        if hvac_mode == HVAC_MODE_HEAT and not inbound:
+            return HVAC_MODE_AUTO
+        elif hvac_mode == HVAC_MODE_AUTO and inbound:
+            return HVAC_MODE_HEAT
+    else:
+        return hvac_mode
+
+
+def calculate_local_setpoint_delta(self) -> Union[float, None]:
+    """Calculate local delta to adjust the setpoint of the TRV based on the air temperature of the external sensor.
+
+    This calibration is for devices with local calibration option, it syncs the current temperature of the TRV to the target temperature of
+    the external sensor.
 
     Parameters
     ----------
@@ -54,240 +84,152 @@ async def startup(self):
 
     Returns
     -------
-    None
+    float
+            new local calibration delta
     """
-    window = None
-    await asyncio.sleep(5)
 
-    while self.startup_running:
-        sensor_state = self.hass.states.get(self.sensor_entity_id)
-        trv_state = self.hass.states.get(self.heater_entity_id)
+    _trv_state_attributes = self.hass.states.get(self.heater_entity_id).attributes
+    _calibration_state = self.hass.states.get(
+        self.local_temperature_calibration_entity
+    ).state
+    _context = "calculate_local_setpoint_delta()"
 
-        if sensor_state is None:
-            _LOGGER.error(
-                "better_thermostat %s: configured temperature sensor with id '%s' could not be located",
-                self.name,
-                self.sensor_entity_id,
-            )
-            return False
-        if trv_state is None:
-            _LOGGER.error(
-                "better_thermostat %s: configured TRV/climate entry with id '%s' could not be located",
-                self.name,
-                self.heater_entity_id,
-            )
-            return False
-        if self.window_id:
-            window = self.hass.states.get(self.window_id)
+    _current_trv_temp = convert_to_float(
+        _trv_state_attributes.get("current_temperature"), self.name, _context
+    )
+    _current_trv_calibration = convert_to_float(_calibration_state, self.name, _context)
 
-            if window is None:
-                _LOGGER.error(
-                    "better_thermostat %s: configured window sensor entry or group with id '%s' could not be located",
-                    self.name,
-                    self.window_id,
-                )
-                return False
-
-            # make sure window has a state variable
-            try:
-                if window.state:
-                    pass
-            except (ValueError, NameError, AttributeError):
-                _LOGGER.error(
-                    "better_thermostat %s: configured window sensor entry or group with id '%s' could not be located",
-                    self.name,
-                    self.window_id,
-                )
-                return False
-
-        _ready = True
-
-        if sensor_state.state in (STATE_UNAVAILABLE, STATE_UNKNOWN, None):
-            _LOGGER.info(
-                "better_thermostat %s: waiting for sensor entity with id '%s' to become fully available...",
-                self.name,
-                self.sensor_entity_id,
-            )
-            _ready = False
-        if trv_state.state in (STATE_UNAVAILABLE, STATE_UNKNOWN, None):
-            _LOGGER.info(
-                "better_thermostat %s: waiting for TRV/climate entity with id '%s' to become fully available...",
-                self.name,
-                self.heater_entity_id,
-            )
-            _ready = False
-        if self.window_id:
-            if self.window_id in (
-                STATE_UNAVAILABLE,
-                STATE_UNKNOWN,
-                None,
-            ) or window.state in (STATE_UNAVAILABLE, STATE_UNKNOWN, None):
-                _LOGGER.info(
-                    "better_thermostat %s: waiting for window sensor entity with id '%s' to become fully available...",
-                    self.name,
-                    self.window_id,
-                )
-                _ready = False
-
-        if not _ready:
-            _LOGGER.info(
-                "better_thermostat %s: retry startup in 15 seconds...", self.name
-            )
-            await asyncio.sleep(15)
-            continue
-
-        if self.window_id:
-            window = self.hass.states.get(self.window_id)
-
-            check = window.state
-            if check == "on":
-                self.window_open = True
-            else:
-                self.window_open = False
-            _LOGGER.debug(
-                "better_thermostat %s: detected window state at startup: %s",
-                self.name,
-                "Open" if self.window_open else "Closed",
-            )
-        if not self.window_id:
-            self.window_open = False
-
-        self.startup_running = False
-        # Use the same precision and min and max as the TRV
-        if (
-            self.hass.states.get(self.heater_entity_id).attributes.get(
-                "target_temp_step"
-            )
-            is not None
-        ):
-            self._TRV_target_temp_step = float(
-                self.hass.states.get(self.heater_entity_id).attributes.get(
-                    "target_temp_step"
-                )
-            )
-        else:
-            self._TRV_target_temp_step = 1
-        if (
-            self.hass.states.get(self.heater_entity_id).attributes.get("min_temp")
-            is not None
-        ):
-            self._TRV_min_temp = float(
-                self.hass.states.get(self.heater_entity_id).attributes.get("min_temp")
-            )
-        else:
-            self._TRV_min_temp = 5
-        if (
-            self.hass.states.get(self.heater_entity_id).attributes.get("max_temp")
-            is not None
-        ):
-            self._TRV_max_temp = float(
-                self.hass.states.get(self.heater_entity_id).attributes.get("max_temp")
-            )
-        else:
-            self._TRV_max_temp = 30
-
-        # Check If we have an old state
-        if (old_state := await self.async_get_last_state()) is not None:
-            # If we have no initial temperature, restore
-            if self._target_temp is None:
-                # If we have a previously saved temperature
-                if old_state.attributes.get(ATTR_TEMPERATURE) is None:
-                    self._target_temp = self._min_temp
-                    _LOGGER.debug(
-                        "better_thermostat %s: Undefined target temperature, falling back to %s",
-                        self.name,
-                        self._target_temp,
-                    )
-                else:
-                    _old_target_temperature = float(
-                        old_state.attributes.get(ATTR_TEMPERATURE)
-                    )
-                    # if the saved temperature is lower than the _min_temp, set it to _min_temp
-                    if _old_target_temperature < self._min_temp:
-                        _LOGGER.warning(
-                            "better_thermostat %s: Saved target temperature %s is lower than _min_temp %s, setting to _min_temp",
-                            self.name,
-                            _old_target_temperature,
-                            self._min_temp,
-                        )
-                        _old_target_temperature = self._min_temp
-                    # if the saved temperature is higher than the _max_temp, set it to _max_temp
-                    elif _old_target_temperature > self._max_temp:
-                        _LOGGER.warning(
-                            "better_thermostat %s: Saved target temperature %s is higher than _max_temp %s, setting to _max_temp",
-                            self.name,
-                            _old_target_temperature,
-                            self._min_temp,
-                        )
-                        _old_target_temperature = self._max_temp
-                    self._target_temp = _old_target_temperature
-            if not self._bt_hvac_mode and old_state.state:
-                self._bt_hvac_mode = old_state.state
-            if not old_state.attributes.get(ATTR_STATE_LAST_CHANGE):
-                self.last_change = old_state.attributes.get(ATTR_STATE_LAST_CHANGE)
-            else:
-                self.last_change = HVAC_MODE_OFF
-            if not old_state.attributes.get(ATTR_STATE_WINDOW_OPEN):
-                self.window_open = old_state.attributes.get(ATTR_STATE_WINDOW_OPEN)
-            if not old_state.attributes.get(ATTR_STATE_DAY_SET_TEMP):
-                self.last_daytime_temp = old_state.attributes.get(
-                    ATTR_STATE_DAY_SET_TEMP
-                )
-            if not old_state.attributes.get(ATTR_STATE_CALL_FOR_HEAT):
-                self.call_for_heat = old_state.attributes.get(ATTR_STATE_CALL_FOR_HEAT)
-            if not old_state.attributes.get(ATTR_STATE_NIGHT_MODE):
-                self.night_mode_active = old_state.attributes.get(ATTR_STATE_NIGHT_MODE)
-                if self.night_mode_active:
-                    if self.night_temp and isinstance(self.night_temp, numbers.Number):
-                        _night_temp = float(self.night_temp)
-                        # if the night temperature is lower than _min_temp, set it to _min_temp
-                        if _night_temp < self._min_temp:
-                            _LOGGER.error(
-                                "better_thermostat %s: Night temperature %s is lower than _min_temp %s, setting to _min_temp",
-                                self.name,
-                                _night_temp,
-                                self._min_temp,
-                            )
-                            self._target_temp = self._min_temp
-                        # if the night temperature is higher than the _max_temp, set it to max_temp
-                        elif _night_temp > self._max_temp:
-                            _LOGGER.warning(
-                                "better_thermostat %s: Night temperature %s is higher than _max_temp %s, setting to _max_temp",
-                                self.name,
-                                _night_temp,
-                                self._min_temp,
-                            )
-                            self._target_temp = self._max_temp
-                    else:
-                        _LOGGER.error(
-                            "better_thermostat %s: Night temp '%s' is not a number",
-                            self.name,
-                            str(self.night_temp),
-                        )
-
-        else:
-            # No previous state, try and restore defaults
-            if self._target_temp is None:
-                _LOGGER.info(
-                    "better_thermostat %s: No previously saved temperature found on startup, turning heat off",
-                    self.name,
-                )
-                self._bt_hvac_mode = HVAC_MODE_OFF
-        # if hvac mode could not be restored, turn heat off
-        if not self._bt_hvac_mode:
-            _LOGGER.warning(
-                "better_thermostat %s: No previously hvac mode found on startup, turn heat off",
-                self.name,
-            )
-            self._bt_hvac_mode = HVAC_MODE_OFF
-        self._last_window_state = self.window_open
-        self._cur_temp = convert_to_float(
-            str(sensor_state.state), self.name, "startup()"
+    if None in (_current_trv_calibration, self._cur_temp, _current_trv_temp):
+        _LOGGER.warning(
+            f"better thermostat {self.name}: Could not calculate local setpoint delta in {_context}:"
+            f" current_trv_calibration: {_current_trv_calibration}, current_trv_temp: {_current_trv_temp}, cur_temp: {self._cur_temp}"
         )
-        self.async_write_ha_state()
-        _LOGGER.info("better_thermostat %s: startup completed.", self.name)
-        await self.control_queue_task.put(self)
-    return True
+        return None
+
+    _new_local_calibration = (
+        self._cur_temp - _current_trv_temp + _current_trv_calibration
+    )
+    return _new_local_calibration
+
+
+def calculate_setpoint_override(self) -> Union[float, None]:
+    """Calculate new setpoint for the TRV based on its own temperature measurement and the air temperature of the external sensor.
+
+    This calibration is for devices with no local calibration option, it syncs the target temperature of the TRV to a new target
+    temperature based on the current temperature of the external sensor.
+
+    Parameters
+    ----------
+    self :
+            self instance of better_thermostat
+
+    Returns
+    -------
+    float
+            new target temp with calibration
+    """
+    state = self.hass.states.get(self.heater_entity_id).attributes
+
+    _context = "calculate_setpoint_override()"
+
+    _current_trv_temp = convert_to_float(
+        state.get("current_temperature"), self.name, _context
+    )
+
+    if None in (self._target_temp, self._cur_temp, _current_trv_temp):
+        return None
+
+    _calibrated_setpoint = round_to_half_degree(
+        self._target_temp - self._cur_temp + _current_trv_temp
+    )
+
+    # check if new setpoint is inside the TRV's range, else set to min or max
+    if _calibrated_setpoint < self._TRV_min_temp:
+        _calibrated_setpoint = self._TRV_min_temp
+    if _calibrated_setpoint > self._TRV_max_temp:
+        _calibrated_setpoint = self._TRV_max_temp
+
+    return _calibrated_setpoint
+
+
+def convert_to_float(
+    value: Union[str, int, float], instance_name: str, context: str
+) -> Union[float, None]:
+    """Convert value to float or print error message.
+
+    Parameters
+    ----------
+    value : str, int, float
+            the value to convert to float
+    instance_name : str
+            the name of the instance thermostat
+    context : str
+            the name of the function which is using this, for printing an error message
+
+    Returns
+    -------
+    float
+            the converted value
+    None
+            If error occurred and cannot convert the value.
+    """
+    if isinstance(value, float):
+        return value
+    else:
+        try:
+            return float(value)
+        except (ValueError, TypeError, AttributeError, KeyError):
+            _LOGGER.debug(
+                f"better thermostat {instance_name}: Could not convert '{value}' to float in {context}"
+            )
+            return None
+
+
+def round_to_half_degree(value: Union[int, float, None]) -> Union[float, int, None]:
+    """Rounds numbers to the nearest n.5/n.0
+
+    Parameters
+    ----------
+    value : int, float
+            input value
+
+    Returns
+    -------
+    float, int
+            either an int, if input was an int, or a float rounded to n.5/n.0
+
+    """
+    if value is None:
+        return None
+    elif isinstance(value, float):
+        return round(value * 2) / 2
+    elif isinstance(value, int):
+        return value
+
+
+def round_to_hundredth_degree(
+    value: Union[int, float, None]
+) -> Union[float, int, None]:
+    """Rounds numbers to the nearest n.nn0
+
+    Parameters
+    ----------
+    value : int, float
+            input value
+
+    Returns
+    -------
+    float, int
+            either an int, if input was an int, or a float rounded to n.nn0
+
+    """
+    if value is None:
+        return None
+    elif isinstance(value, float):
+        return round(value * 100) / 100
+    elif isinstance(value, int):
+        return value
+
 
 
 def check_float(potential_float):
@@ -337,3 +279,41 @@ def convert_time(time_string):
         )
     except ValueError:
         return None
+
+async def get_device_model(self):
+    """Fetches the device model from HA.
+    Parameters
+    ----------
+    self :
+            self instance of better_thermostat
+    Returns
+    -------
+    string
+            the name of the thermostat model
+    """
+    if self.model is None:
+        try:
+            entity_reg = await er.async_get(self.hass)
+            entry = entity_reg.async_get(self.heater_entity_id)
+            dev_reg = await dr.async_get(self.hass)
+            device = dev_reg.async_get(entry.device_id)
+            _LOGGER.debug(f"better_thermostat {self.name}: found device:")
+            _LOGGER.debug(device)
+            try:
+                # Z2M reports the device name as a long string with the actual model name in braces, we need to extract it
+                return re.search("\((.+?)\)", device.model).group(1)
+            except AttributeError:
+                # Other climate integrations might report the model name plainly, need more infos on this
+                return device.model
+        except (
+            RuntimeError,
+            ValueError,
+            AttributeError,
+            KeyError,
+            TypeError,
+            NameError,
+            IndexError,
+        ):
+            return "generic"
+    else:
+        return self.model
