--- custom_components/better_thermostat/models/devices/GS361A_H04.yaml
@@ -1,2 +1,0 @@
-calibration_type: 1
-heat_auto_swapped: False
