--- custom_components/better_thermostat/models/devices/SEA801-Zigbee_SEA802-Zigbee.yaml
@@ -1,4 +1,0 @@
-calibration_type: 0
-calibration_round: true
-has_system_mode: True
-heat_auto_swapped: False
