--- custom_components/better_thermostat/models/devices/SPZB0001.yaml
@@ -1,4 +1,0 @@
-calibration_type: 0
-calibration_round: false
-has_system_mode: True
-heat_auto_swapped: True
