--- custom_components/better_thermostat/models/devices/TS0601_thermostat.yaml
@@ -1,3 +1,0 @@
-calibration_type: 1
-has_system_mode: True
-heat_auto_swapped: True
