--- custom_components/better_thermostat/models/devices/TV02-Zigbee.yaml
@@ -1,3 +1,0 @@
-calibration_type: 1
-has_system_mode: True
-heat_auto_swapped: False
