+++ custom_components/better_thermostat/translations/en.json
@@ -1,32 +1,54 @@
 {
-    "title":"Better Thermostat",
     "config": {
-            "flow_title": "Discovered Thermostat ({name})",
             "step": {
                     "user": {
-                            "description": "Set up your Better Thermostat to integrate with Home Assistant.",
+                            "title": "Better Thermostat",
+                            "description": "Setup your Better Thermostat to integrate with Home Assistant",
                             "data": {
-                                "name": "Name",
-                                "thermostat": "The real Themostat",
-                                "temperature_sensor": "Temperature Sensor",
-                                "window_sensors": "Window Sensor",
+                                "name": "Name*",
+                                "thermostat": "The real themostat*",
+                                "temperature_sensor": "Temperature sensor*",
+                                "window_sensors": "Window sensor",
                                 "off_temperature": "The outdoor temperature when the thermostat turn off",
                                 "window_off_delay": "Delay before the thermostat turn off when the window is open and on when the window is closed",
                                 "outdoor_sensor": "If you have an outdoor sensor, you can use it to get the outdoor temperature",
                                 "valve_maintenance": "If your thermostat has no own maintenance mode, you can use this one",
                                 "local_calibration": "The local calibration entity of your thermostat",
-                                "weather": "Your weather entity to get the outdoor temperature"
+                                "weather": "Your weather entity to get the outdoor temperature",
+                                "calibration_round": "Should the calibration be rounded to the nearest",
+                                "heat_auto_swapped": "If the auto means heat for your TRV and you want to swap it (for Google Home users)",
+                                "child_lock": "Ignore all inputs on the TRV like a child lock"
                             }
                     }
             },
             "error": {
                     "failed": "something went wrong.",
-                    "no_calibration_found": "No calibration entity found for this thermostat",
                     "no_outside_temp": "You need to define an outside temperature sensor or weather entity."
             },
             "abort": {
                     "single_instance_allowed": "Only a single Thermostat for each real is allowed.",
                     "no_devices_found": "No thermostat entity found, make sure you have a climate entity in your home assistant"
             }
+    },
+    "options": {
+        "step": {
+                "user": {
+                        "title": "Better Thermostat",
+                        "description": "Update your Better Thermostat settings",
+                        "data": {
+                            "temperature_sensor": "Temperature Sensor",
+                            "window_sensors": "Window Sensor",
+                            "off_temperature": "The outdoor temperature when the thermostat turn off",
+                            "window_off_delay": "Delay before the thermostat turn off when the window is open and on when the window is closed",
+                            "outdoor_sensor": "If you have an outdoor sensor, you can use it to get the outdoor temperature",
+                            "valve_maintenance": "If your thermostat has no own maintenance mode, you can use this one",
+                            "local_calibration": "The local calibration entity of your thermostat",
+                            "weather": "Your weather entity to get the outdoor temperature",
+                            "calibration_round": "Should the calibration be rounded to the nearest",
+                            "heat_auto_swapped": "If the auto means heat for your TRV and you want to swap it",
+                            "child_lock": "Ignore all inputs on the TRV like a child lock"
+                        }
+                }
+        }
     }
 }
