+++ custom_components/better_thermostat/weather.py
@@ -6,7 +6,7 @@ from datetime import datetime, timedelta
 import homeassistant.util.dt as dt_util
 from homeassistant.components.recorder import history
 
-from .models.utils import convert_to_float
+from .helpers import convert_to_float
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -34,7 +34,7 @@ def check_weather(self) -> bool | None:
         _LOGGER.debug(
             f"better_thermostat {self.name}: checking ambient air sensor data..."
         )
-        self.call_for_heat = check_ambient_air_temperature(self)
+        self.call_for_heat = self._hass.async_add_executor_job(check_ambient_air_temperature(self))
     else:
         _LOGGER.debug(
             f"better_thermostat {self.name}: could not find any weather sensors... setting call_for_heat to true"
