+++ hacs.json
@@ -1,5 +1,5 @@
 {
   "name": "Better Thermostat",
   "render_readme": true,
-  "homeassistant": "2021.12.0"
+  "homeassistant": "2022.9.0"
 }
